Refactor MarketSymbol: consolidate into a single file and enhance contract type representation
